requirements.txt
django==5.0
djangorestframework==3.14.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
spacy==3.8.8
word2number==1.1

